moves code; conditionaly renders analytics for guest and supervisor
