+++ src/components/pages/guest-pages/GuestAnalytics.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const GuestAnalytics = () => {
+  return (
+    <div>
+      <h1> Guest Analytics</h1>
+    </div>
+  );
+};
+
+export default GuestAnalytics;
