+++ src/components/pages/guest-pages/GuestDashboard.jsx
@@ -4,6 +4,9 @@ const GuestDashboard = () => {
   return (
     <div className="container">
       <h1>Guest dashboard</h1>
+      <h1>Welcome To Family Promise of Spokane</h1>
+      <h2>The facility has 12 beds left.</h2>
+      <p> Would you like to check in? </p>
     </div>
   );
 };
